+++ changelog.md
@@ -1,3 +1,7 @@
+# 12.2 NetworkClient fix
+- отрефакторил NetworkClientDio, чтобы не было повторяющегося кода. Добавил функ. _executeRequest.
+- Для развязки с dio сделал интерфейс NetworkClient от которого теперь зависит SightRepositoryNetwork. И при создании SightRepositoryNetwork в него инжектится нужная имплементация, типа NetworkClientDio или для опытов NetworkClientNoInternet. Можно в целом теперь реализовать на пакете http, к примеру. Или для тестов, моковый клиент.
+
 # 12.2 NetworkClient
 - добавил NetworkClient, который совершает запросы с серверу и в случае успеха возвращает ответ или выкидывает соответствующий эксепшен. Логика взаимодействия с сервером и ошибок инкапсулирова в этом классе.
