+++ lib/config.dart
@@ -2,4 +2,7 @@ class Config {
   /// Значения для фильтра.  min max радиусы поиска
   static const double minRange = 100.0;
   static const double maxRange = 100000.0;
+
+  /// URL сервера
+  static const String baseUrl = 'https://test-backend-flutter.surfstudio.ru';
 }
