+++ lib/main.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:places/config.dart';
 import 'package:places/data/favorites_repository/favorites_repository_memory.dart';
 import 'package:places/data/location_repository/location_repository_mock.dart';
+import 'package:places/data/network_client/network_client.dart';
+import 'package:places/data/network_client/network_client_dio.dart';
 import 'package:places/data/sight_repository/sight_repository_network.dart';
 import 'package:places/data/visited_repository/visited_repository_memory.dart';
 import 'package:places/domain/filter.dart';
@@ -31,9 +34,14 @@ import 'package:places/ui/screen/visiting_screen.dart';
 // Тут, пока не внедряли других решений
 final searchHistoryState = SearchHistoryState();
 
-// Временное место для интеракторов
+// Временное место для интеракторов, репозиториев и тп
 // final SightRepository sightRepository = SightRepositoryMemory();
-final SightRepository sightRepository = SightRepositoryNetwork();
+
+final NetworkClient networkClient = NetworkClientDio(baseUrl: Config.baseUrl);
+// final NetworkClient networkClient = NetworkClientNoInternet();
+
+final SightRepository sightRepository = SightRepositoryNetwork(networkClient);
+
 final FavoritesRepository favoritesRepository = FavoritesRepositoryMemory();
 final VisitedRepository visitedRepository = VisitedRepositoryMemory();
 final LocationRepository locationRepository = LocationRepositoryMock();
